php: comply with psr-4 autoloading
