+++ php/init.php
@@ -1,4 +1,4 @@
 <?php
 
 require __DIR__ . '/src/Webhook.php';
-require __DIR__ . '/src/Exception/WebhookException.php';
+require __DIR__ . '/src/Exception/WebhookVerificationException.php';
